+++ packages/app-config-writer/test/unit/prompt/smartlinks-config.test.ts
@@ -52,6 +52,7 @@ describe('Test function getSmartLinksTargetFromPrompt', () => {
 
     beforeEach(() => {
         jest.resetAllMocks();
+        jest.restoreAllMocks();
         isAppStudioMock = jest.spyOn(btp, 'isAppStudio');
         getSystemCredentialsSpy = jest.spyOn(utils, 'getLocalStoredCredentials');
         listDestinationsMock = jest.spyOn(btp, 'listDestinations');