+++ packages/axios-extension/src/factory.ts
@@ -153,6 +153,7 @@ export function createForAbapOnCloud(options: AbapCloudOptions & Partial<Provide
             const { url, cookies, ...config } = options;
             provider = createInstance<AbapServiceProvider>(AbapServiceProvider, {
                 baseURL: url,
+                cookies,
                 ...config
             });
             if (!cookies) {