+++ packages/deploy-tooling/test/unit/base/deploy.test.ts
@@ -17,7 +17,6 @@ import * as validate from '../../../src/base/validate';
 import { SummaryStatus } from '../../../src/base/validate';
 
 const validateBeforeDeployMock = jest.spyOn(validate, 'validateBeforeDeploy');
-const formatSummaryMock = jest.spyOn(validate, 'formatSummary');
 const showAdditionalInfoForOnPremMock = jest.spyOn(validate, 'showAdditionalInfoForOnPrem');
 
 describe('base/deploy', () => {
@@ -48,7 +47,8 @@ describe('base/deploy', () => {
             mockedLrepService.deploy.mockReset();
             mockedAdtService.createTransportRequest.mockReset();
             validateBeforeDeployMock.mockReset();
-            formatSummaryMock.mockReset();
+            jest.clearAllMocks();
+            jest.restoreAllMocks();
         });
 
         test('No errors locally with url', async () => {
@@ -90,6 +90,7 @@ describe('base/deploy', () => {
         });
 
         test('Log validation summaries regardless of validation result', async () => {
+            const formatSummaryMock = jest.spyOn(validate, 'formatSummary');
             mockedStoreService.read.mockResolvedValueOnce(credentials);
             mockedUi5RepoService.deploy.mockResolvedValue(undefined);
             validateBeforeDeployMock.mockResolvedValueOnce({
@@ -275,7 +276,7 @@ describe('base/deploy', () => {
             jest.spyOn(nullLogger, 'info');
             showAdditionalInfoForOnPremMock.mockResolvedValue(true);
             await deploy(archive, { app, target }, nullLogger);
-            expect(nullLogger.info).toHaveBeenCalledTimes(3);
+            expect(nullLogger.info).toHaveBeenCalledTimes(2);
         });
 
         describe('adaptation projects', () => {