+++ packages/fe-fpm-writer/src/action/index.ts
@@ -45,6 +45,7 @@ function enhanceConfig(data: CustomAction, manifestPath: string, manifest: Manif
  */
 export function enhanceManifestAndGetActionsElementReference(manifest: any, target: CustomActionTarget): any {
     const page = manifest['sap.ui5'].routing.targets[target.page];
+    const SECTIONS = 'sections';
     page.options = page.options || {};
     page.options.settings = page.options.settings || {};
     if (target.control === TargetControl.header || target.control === TargetControl.footer) {
@@ -53,6 +54,15 @@ export function enhanceManifestAndGetActionsElementReference(manifest: any, targ
         page.options.settings.content[target.control].actions =
             page.options.settings.content[target.control].actions || {};
         return page.options.settings.content[target.control].actions;
+    } else if (target.control === TargetControl.body && target.customSectionKey) {
+        // custom section actions
+        page.options.settings[target.control] = page.options.settings[target.control] || {};
+        page.options.settings[target.control][SECTIONS] = page.options.settings[target.control][SECTIONS] || {};
+        page.options.settings[target.control][SECTIONS][target.customSectionKey] =
+            page.options.settings[target.control][SECTIONS][target.customSectionKey] || {};
+        page.options.settings[target.control][SECTIONS][target.customSectionKey].actions =
+            page.options.settings[target.control][SECTIONS][target.customSectionKey].actions || {};
+        return page.options.settings[target.control][SECTIONS][target.customSectionKey].actions;
     } else {
         const controlPrefix = target.navProperty ? target.navProperty + '/' : '';
         const controlSuffix = target.qualifier ? '#' + target.qualifier : '';