+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.7.33
+
+### Patch Changes
+
+-   Updated dependencies [558891c2]
+    -   @sap-ux/axios-extension@1.12.6
+
+## 0.7.32
+
+### Patch Changes
+
+-   Updated dependencies [69282b7d]
+    -   @sap-ux/axios-extension@1.12.5
+
 ## 0.7.31
 
 ### Patch Changes