+++ packages/cards-editor-middleware/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/cards-editor-middleware
 
+## 0.3.9
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.20.3
+
+## 0.3.8
+
+### Patch Changes
+
+-   Updated dependencies [54c91c6d]
+    -   @sap-ux/project-access@1.20.2
+
 ## 0.3.7
 
 ### Patch Changes