+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/project-access
 
+## 1.20.3
+
+### Patch Changes
+
+-   Updated dependencies [f80a4256]
+    -   @sap-ux/i18n@0.0.7
+
+## 1.20.2
+
+### Patch Changes
+
+-   54c91c6d: Export method 'clearCdsModuleCache' to provide option to clear cds module cache for passed project root path.
+
 ## 1.20.1
 
 ### Patch Changes